requirements.txt
-i https://pypi.org/simple
aniso8601==9.0.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
crontab==1.0.1
django==4.1
django-annoying==0.10.6
django-crontab==0.7.1
django-filter==22.1
django-leaflet==0.29.0
djangorestframework==3.14.0
geographiclib==2.0; python_version >= '3.7'
geojson==3.0.1
geopy==2.3.0
graphene==3.2.2
graphene-django==3.1.1
graphene-django-extras==1.0.0
graphene-gis==0.0.8
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.3
osm2geojson==0.2.4
pillow==9.5.0
promise==2.3
pyproj==3.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
setuptools==68.0.0; python_version >= '3.7'
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.6.3
tzlocal==5.0.1; python_version >= '3.7'
tzdata==2023.3
urllib3==2.0.3
